requirements.txt
3es
docker==5.0.3
dvc==2.10.2
dvc-render==0.0.5
flask
flask_cors
gevent
gitdb==4.0.9
GitPython==3.1.27
google-api-core==2.7.3
google-api-python-client==2.48.0
google-auth==2.6.6
google-auth-httplib2==0.1.0
googleapis-common-protos==1.56.1
isort==5.10.1
mlflow==1.26.0
mysql-connector==2.2.9
numpy==1.22.3
oauth2client
plotly==5.7.0
py==1.11.0
pydot==1.4.2
PyDrive==1.3.1
PyDrive2==1.10.1
pytest==7.1.2
regex==2022.4.24
requests==2.26.0
requests-oauthlib==1.3.0
ruamel.yaml.clib==0.2.6
s3transfer==0.5.0
scikit-learn==1.0.2
scipy==1.8.0
scmrepo==0.0.19
seaborn==0.11.2
semver==2.13.0
setuptools
sklearn==0.0
tqdm==4.64.0
traitlets==5.1.1
matplotlib
